TYP: Add the `np.exceptions` namespace and add `DTypePromotionError`
Xref https://github.com/numpy/numpy/pull/22644 and https://github.com/numpy/numpy/pull/22707

Update modules.pyi
