+++ numpy/typing/tests/data/reveal/modules.pyi
@@ -16,6 +16,7 @@ reveal_type(np.random)  # E: ModuleType
 reveal_type(np.rec)  # E: ModuleType
 reveal_type(np.testing)  # E: ModuleType
 reveal_type(np.version)  # E: ModuleType
+reveal_type(np.exceptions)  # E: ModuleType
 
 reveal_type(np.lib.format)  # E: ModuleType
 reveal_type(np.lib.mixins)  # E: ModuleType
